Move KnowledgeProvider
